Move feedback into components/
 🐿 v2.8.0
